refactor: unify storage block and disk drive network nodes
