+++ refinedstorage2-network-test/src/main/java/com/refinedmods/refinedstorage2/network/test/nodefactory/StorageNetworkNodeFactory.java
@@ -7,8 +7,17 @@ import com.refinedmods.refinedstorage2.network.test.AddNetworkNode;
 import java.util.Map;
 
 public class StorageNetworkNodeFactory extends AbstractNetworkNodeFactory {
+    public static final String PROPERTY_ENERGY_USAGE_PER_STORAGE = "energy_usage_per_storage";
+    public static final String PROPERTY_SIZE = "size";
+
     @Override
     protected AbstractNetworkNode innerCreate(final AddNetworkNode ctx, final Map<String, Object> properties) {
-        return new StorageNetworkNode(getEnergyUsage(properties));
+        final long energyUsagePerStorage = (long) properties.getOrDefault(PROPERTY_ENERGY_USAGE_PER_STORAGE, 0L);
+        final int size = (int) properties.getOrDefault(PROPERTY_SIZE, 9);
+        return new StorageNetworkNode(
+            getEnergyUsage(properties),
+            energyUsagePerStorage,
+            size
+        );
     }
 }
