+++ refinedstorage2-network/src/test/java/com/refinedmods/refinedstorage2/api/network/impl/node/storage/StorageNetworkNodeProviderImpl.java
@@ -1,4 +1,4 @@
-package com.refinedmods.refinedstorage2.api.network.impl.node.multistorage;
+package com.refinedmods.refinedstorage2.api.network.impl.node.storage;
 
 import com.refinedmods.refinedstorage2.api.storage.Storage;
 
@@ -6,7 +6,7 @@ import java.util.HashMap;
 import java.util.Map;
 import java.util.Optional;
 
-class MultiStorageProviderImpl implements MultiStorageProvider {
+class StorageNetworkNodeProviderImpl implements StorageNetworkNode.Provider {
     private final Map<Integer, Storage> storages = new HashMap<>();
 
     @Override
